+++ src/kafunk/Consumer.fs
@@ -365,6 +365,10 @@ type ConsumerMessageSet =
     /// Returns the last offset in the message set.
     static member lastOffset (ms:ConsumerMessageSet) =
       MessageSet.lastOffset ms.messageSet
+
+    /// Returns the timestamp of the first message.
+    static member firstTimestamp (ms:ConsumerMessageSet) =
+      MessageSet.firstTimestamp ms.messageSet
     
     /// Returns the offset to commit when this message set has been consumed.
     static member commitOffset (ms:ConsumerMessageSet) =
@@ -593,7 +597,7 @@ module Consumer =
     let cfg = c.config
     let topic = cfg.topic
     let fetch = Kafka.fetch c.conn |> AsyncFunc.catch
-    let messageVer = Versions.fetchResMessage (Versions.byKey c.conn.Config.version ApiKey.Fetch)
+    let messageVer = MessageVersions.fetchResMessage (c.conn.ApiVersion ApiKey.Fetch)
     Group.tryAsync
       (state)
       (fun _ -> None)