+++ src/kafunk/Protocol.fs
@@ -5,58 +5,26 @@ namespace Kafunk
 module Protocol =
 
   type ApiKey =
-    | Produce = 0s
-    | Fetch = 1s
-    | Offset = 2s
-    | Metadata = 3s
-    | OffsetCommit = 8s
-    | OffsetFetch = 9s
-    | GroupCoordinator = 10s
-    | JoinGroup = 11s
-    | Heartbeat = 12s
-    | LeaveGroup = 13s
-    | SyncGroup = 14s
-    | DescribeGroups = 15s
-    | ListGroups = 16s
+    | Produce = 0
+    | Fetch = 1
+    | Offset = 2
+    | Metadata = 3
+    | OffsetCommit = 8
+    | OffsetFetch = 9
+    | GroupCoordinator = 10
+    | JoinGroup = 11
+    | Heartbeat = 12
+    | LeaveGroup = 13
+    | SyncGroup = 14
+    | DescribeGroups = 15
+    | ListGroups = 16
+    | ApiVersions = 18
 
   type ApiVersion = int16
 
   [<Compile(Module)>]
-  module Versions =
+  module internal MessageVersions =
     
-    let V_0_8_2 = System.Version (0, 8, 2)
-    let V_0_9_0 = System.Version (0, 9, 0)
-    let V_0_10_0 = System.Version (0, 10, 0)
-    let V_0_10_1 = System.Version (0, 10, 1)
-
-    /// Returns an ApiVersion given a system version and an ApiKey.
-    let internal byKey (version:System.Version) (apiKey:ApiKey) : ApiVersion = 
-      match apiKey with
-      | ApiKey.OffsetFetch -> 
-        if version >= V_0_9_0 then 1s
-        elif version >= V_0_8_2 then 0s
-        else failwith "not supported"
-      | ApiKey.OffsetCommit -> 
-        if version >= V_0_9_0 then 2s
-        elif version >= V_0_8_2 then 1s
-        else 0s
-      | ApiKey.Produce -> 
-        if version >= V_0_10_0 then 2s
-        elif version >= V_0_9_0 then 1s
-        else 0s
-      | ApiKey.Fetch ->
-        if version >= V_0_10_0 then 2s
-        elif version >= V_0_9_0 then 1s
-        else 0s
-      | ApiKey.JoinGroup -> 
-        if version >= V_0_10_1 then 1s
-        else 0s
-      | ApiKey.Offset ->
-        if version >= V_0_10_1 then 1s
-        else 0s
-      | _ -> 
-        0s
-
     /// Gets the version of Message for a ProduceRequest of the specified API version.
     let internal produceReqMessage (apiVer:ApiVersion) =
       if apiVer >= 2s then 1s
@@ -643,7 +611,7 @@ module Protocol =
       size
 
     static member internal Write (ver:ApiVersion, x:ProduceRequest, buf:BinaryZipper) =
-      let messageVer = Versions.produceReqMessage ver
+      let messageVer = MessageVersions.produceReqMessage ver
       buf.WriteInt16 x.requiredAcks
       buf.WriteInt32 x.timeout
       buf.WriteInt32 x.topics.Length
@@ -772,7 +740,7 @@ module Protocol =
           let errorCode = buf.ReadInt16 ()
           let hwo = buf.ReadInt64 ()
           let mss = buf.ReadInt32 ()
-          let ms = MessageSet.Read (Versions.fetchResMessage ver,partition,errorCode,mss,buf)
+          let ms = MessageSet.Read (MessageVersions.fetchResMessage ver,partition,errorCode,mss,buf)
           ps.[j] <-  partition, errorCode, hwo, mss, ms
         topics.[i] <- (t,ps)
       let res = FetchResponse(throttleTime, topics)
@@ -837,13 +805,15 @@ module Protocol =
         let ts = 
           if ver >= 1s then buf.ReadInt64 ()
           else 0L
-        let os = buf.ReadArray (fun buf -> buf.ReadInt64())
+        let os = 
+          if ver >= 1s then [|buf.ReadInt64 ()|]
+          else buf.ReadArray (fun buf -> buf.ReadInt64 ())
         PartitionOffsets(p, ec, ts, os)
       let readTopic (buf:BinaryZipper) =
         let t = buf.ReadString ()
         let ps = buf.ReadArray readPartition
         t,ps
-      let topics = buf.ReadArray (readTopic)
+      let topics = buf.ReadArray readTopic
       OffsetResponse(topics)
 
   // Offset Commit/Fetch API
@@ -1070,9 +1040,10 @@ module Protocol =
         { errorCode = errorCode; generationId = generationId; groupProtocol = groupProtocol;
           leaderId = leaderId; memberId = memberId; members = members }
 
-    let internal sizeRequest (req:Request) =
+    let internal sizeRequest (ver:ApiVersion, req:Request) =
       Binary.sizeString req.groupId +
       Binary.sizeInt32 req.sessionTimeout +
+      (if ver >= 1s then 4 else 0) +
       Binary.sizeString req.memberId +
       Binary.sizeString req.protocolType +
       GroupProtocols.size req.groupProtocols
@@ -1396,6 +1367,34 @@ module Protocol =
       let xs, buf = buf |> Binary.readArray readGroup
       (DescribeGroupsResponse(xs), buf)
 
+  [<NoEquality;NoComparison>]
+  type ApiVersionsRequest =
+    struct end
+    with
+      static member Size (_:ApiVersionsRequest) = 0
+      static member internal Write (_:ApiVersionsRequest, _:BinaryZipper) = ()
+
+  type MinVersion = int16
+  type MaxVersion = int16
+
+  [<NoEquality;NoComparison>]
+  type ApiVersionsResponse =
+    struct
+      val errorCode : ErrorCode
+      val apiVersions : (ApiKey * MinVersion * MaxVersion)[]
+      new (ec,apiVersions) = { errorCode = ec ; apiVersions = apiVersions }
+    end
+    with
+      static member internal Read (_:ApiVersion, buf:BinaryZipper) =
+        let ec = buf.ReadInt16 ()
+        let apiVersions = buf.ReadArray (fun buf ->
+          let apiKey : ApiKey = enum<ApiKey> (int (buf.ReadInt16 ()))
+          let min = buf.ReadInt16 ()
+          let max = buf.ReadInt16 ()
+          apiKey,min,max)
+        ApiVersionsResponse(ec,apiVersions)
+        
+
   /// A Kafka request message.
   type RequestMessage =
     | Metadata of Metadata.Request
@@ -1411,6 +1410,7 @@ module Protocol =
     | LeaveGroup of LeaveGroupRequest
     | ListGroups of ListGroupsRequest
     | DescribeGroups of DescribeGroupsRequest
+    | ApiVersions of ApiVersionsRequest
   with
 
     static member internal size (ver:ApiVersion, x:RequestMessage) =
@@ -1423,11 +1423,12 @@ module Protocol =
       | GroupCoordinator x -> GroupCoordinatorRequest.size x
       | OffsetCommit x -> OffsetCommitRequest.Size (ver,x)
       | OffsetFetch x -> OffsetFetchRequest.size x
-      | JoinGroup x -> JoinGroup.sizeRequest x
+      | JoinGroup x -> JoinGroup.sizeRequest (ver,x)
       | SyncGroup x -> SyncGroupRequest.size x
       | LeaveGroup x -> LeaveGroupRequest.size x
       | ListGroups x -> ListGroupsRequest.size x
       | DescribeGroups x -> DescribeGroupsRequest.size x
+      | ApiVersions x -> ApiVersionsRequest.Size x
 
     static member internal Write (ver:ApiVersion, x:RequestMessage, buf:BinaryZipper) =
       match x with
@@ -1444,6 +1445,7 @@ module Protocol =
       | LeaveGroup x -> LeaveGroupRequest.write x buf.Buffer |> ignore
       | ListGroups x -> ListGroupsRequest.write x buf.Buffer |> ignore
       | DescribeGroups x -> DescribeGroupsRequest.write x buf.Buffer |> ignore
+      | ApiVersions x -> ApiVersionsRequest.Write (x,buf)
 
     member x.ApiKey =
       match x with
@@ -1460,6 +1462,7 @@ module Protocol =
       | LeaveGroup _ -> ApiKey.LeaveGroup
       | ListGroups _ -> ApiKey.ListGroups
       | DescribeGroups _ -> ApiKey.DescribeGroups
+      | ApiVersions _ -> ApiKey.ApiVersions
 
   /// A Kafka request envelope.
   type Request =
@@ -1504,6 +1507,7 @@ module Protocol =
     | LeaveGroupResponse of LeaveGroupResponse
     | ListGroupsResponse of ListGroupsResponse
     | DescribeGroupsResponse of DescribeGroupsResponse
+    | ApiVersionsResponse of ApiVersionsResponse
   with
 
     /// Decodes the response given the specified ApiKey corresponding to the request.
@@ -1532,7 +1536,9 @@ module Protocol =
         let x, _ = ListGroupsResponse.read buf.Buffer in (ResponseMessage.ListGroupsResponse x)
       | ApiKey.DescribeGroups ->
         let x, _ = DescribeGroupsResponse.read buf.Buffer in (ResponseMessage.DescribeGroupsResponse x)
-      | x -> failwith (sprintf "Unsupported ApiKey=%A" x)
+      | ApiKey.ApiVersions -> ApiVersionsResponse.Read (apiVer,buf) |> ResponseMessage.ApiVersionsResponse
+      | x -> 
+        failwith (sprintf "Unsupported ApiKey=%A" x)
 
   /// A Kafka response envelope.
   type Response =